fix heartbeat tests
